backend/src/requirements.txt
aiomysql==0.1.1
aiopg==1.4.0
aiosqlite==0.19.0
anyio==3.6.2
async-generator==1.10
async-timeout==4.0.2
asyncmy==0.2.7
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
databases==0.7.0
exceptiongroup==1.1.1
fastapi==0.95.0
geographiclib==2.0
geopy==2.3.0
greenlet==2.0.2
gtin-validator==1.0.3
h11==0.14.0
idna==3.4
iniconfig==2.0.0
outcome==1.2.0
packaging==23.1
pluggy==1.0.0
psycopg2-binary==2.9.6
pydantic==1.10.7
PyMySQL==1.0.3
PySocks==1.7.1
pytest==7.3.1
requests==2.28.2
selenium==4.9.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.47
starlette==0.26.1
tomli==2.0.1
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
wsproto==1.2.0
